+++ src/components/stepforms/SortStepForm.vue
@@ -1,7 +1,6 @@
 <template>
   <div>
-    <step-form-title :title="title" :stepName="this.editedStep.name" />
-
+    <step-form-header :cancel="cancelEdition" :title="title" :stepName="this.editedStep.name" />
     <ListWidget
       addFieldName="Add Column"
       id="sortColumn"