+++ tests/unit/aggregate-step-form.spec.ts
@@ -8,7 +8,6 @@ import { Pipeline } from '@/lib/steps';
 
 import { setupMockStore, RootState } from './utils';
 
-
 const localVue = createLocalVue();
 localVue.use(Vuex);
 
@@ -248,7 +247,7 @@ describe('Aggregate Step Form', () => {
 
   it('should emit "cancel" event when edition is cancelled', () => {
     const wrapper = mount(AggregateStepForm, { store: emptyStore, localVue, sync: false });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(wrapper.emitted()).toEqual({
       cancel: [[]],
     });
@@ -280,10 +279,10 @@ describe('Aggregate Step Form', () => {
       propsData: { isStepCreation: true },
       sync: false,
     });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(2);
     wrapper.setProps({ isStepCreation: false });
-    await wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    await wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(3);
   });
 });