+++ tests/unit/append-step-form.spec.ts
@@ -6,7 +6,6 @@ import { Pipeline } from '@/lib/steps';
 
 import { setupMockStore, RootState } from './utils';
 
-
 const localVue = createLocalVue();
 localVue.use(Vuex);
 
@@ -82,7 +81,7 @@ describe('Append Step Form', () => {
     });
 
     const wrapper = mount(AppendStepForm, { store, localVue });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(wrapper.emitted()).toEqual({ cancel: [[]] });
     expect(store.state.vqb.selectedStepIndex).toEqual(1);
     expect(store.state.vqb.pipeline).toEqual([
@@ -104,10 +103,10 @@ describe('Append Step Form', () => {
     });
     const wrapper = mount(AppendStepForm, { store, localVue });
     wrapper.setProps({ isStepCreation: true });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(2);
     wrapper.setProps({ isStepCreation: false });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(3);
   });
 });