+++ tests/unit/percentage-step-form.spec.ts
@@ -76,7 +76,7 @@ describe('Percentage Step Form', () => {
 
   it('should emit "cancel" event when edition is cancelled', async () => {
     const wrapper = mount(PercentageStepForm, { store: emptyStore, localVue });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     await localVue.nextTick();
     expect(wrapper.emitted()).toEqual({
       cancel: [[]],
@@ -96,10 +96,10 @@ describe('Percentage Step Form', () => {
     });
     const wrapper = mount(PercentageStepForm, { store, localVue });
     wrapper.setProps({ isStepCreation: true });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(2);
     wrapper.setProps({ isStepCreation: false });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(3);
   });