+++ tests/unit/rename-step-form.spec.ts
@@ -101,7 +101,7 @@ describe('Rename Step Form', () => {
 
   it('should emit "cancel" event when edition is cancelled', async () => {
     const wrapper = mount(RenameStepForm, { store: emptyStore, localVue });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     await localVue.nextTick();
     expect(wrapper.emitted()).toEqual({
       cancel: [[]],
@@ -135,10 +135,10 @@ describe('Rename Step Form', () => {
     });
     const wrapper = mount(RenameStepForm, { store, localVue });
     wrapper.setProps({ isStepCreation: true });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(2);
     wrapper.setProps({ isStepCreation: false });
-    wrapper.find('.widget-form-action__button--cancel').trigger('click');
+    wrapper.find('.step-edit-form__back-button').trigger('click');
     expect(store.state.vqb.selectedStepIndex).toEqual(3);
   });